pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = "Demonstrtion code base for HR"
authors = ["stix <stix_simfi@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.4.2"
allure-pytest = "^2.13.5"
locust = "^2.32.2"
pytest-xdist = "3.6.1"
flake8 = "^7.1.1"
selenium = "^4.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"